Refactored db such that the seeding file is seperate from index.js
